+++ sqflite/lib/sqflite.dart
@@ -11,8 +11,6 @@ import 'sqlite_api.dart';
 
 export 'package:sqflite/sql.dart' show ConflictAlgorithm;
 export 'package:sqflite/src/compat.dart';
-export 'package:sqflite/src/constant.dart'
-    show sqfliteLogLevelNone, sqfliteLogLevelSql, sqfliteLogLevelVerbose;
 export 'package:sqflite/src/factory_impl.dart' show databaseFactory;
 
 export 'sqlite_api.dart';
@@ -21,13 +19,6 @@ export 'sqlite_api.dart';
 /// sqflite plugin
 ///
 class Sqflite {
-  //static MethodChannel get _channel => channel;
-
-  /// deprecated
-  @deprecated
-  static Future<String> get platformVersion =>
-      invokeMethod<String>(methodGetPlatformVersion);
-
   /// Turns on debug mode if you want to see the SQL query
   /// executed natively.
   static Future<void> setDebugModeOn([bool on = true]) async {
@@ -49,16 +40,16 @@ class Sqflite {
     return setDebugModeOn(on);
   }
 
-  // Testing only
+  /// Testing only.
+  ///
   /// deprecated on purpose to remove from code.
   @deprecated
   static Future<void> devSetOptions(SqfliteOptions options) async {
     await invokeMethod<dynamic>(methodOptions, options.toMap());
   }
 
-  @deprecated
-
   /// Testing only
+  @deprecated
   static Future<void> devInvokeMethod(String method,
       [dynamic arguments]) async {
     await invokeMethod<dynamic>(method, arguments);
@@ -69,8 +60,8 @@ class Sqflite {
   static int firstIntValue(List<Map<String, dynamic>> list) =>
       utils.firstIntValue(list);
 
-  /// Utility to encode a blob to allow blow query using
-  /// "hex(blob_field) = ?", Sqlite.hex([1,2,3])
+  /// Utility to encode a blob to allow blob query using
+  /// 'hex(blob_field) = ?', Sqlite.hex([1,2,3])
   static String hex(List<int> bytes) => utils.hex(bytes);
 
   /// Sqlite has a dead lock warning feature that will print some text
@@ -150,9 +141,9 @@ Future<Database> openDatabase(String path,
     OnDatabaseVersionChangeFn onDowngrade,
     OnDatabaseOpenFn onOpen,
     bool readOnly = false,
-    bool singleInstance = true,
-    int cursorWindowSize}) {
-  final OpenDatabaseOptions options = OpenDatabaseOptions(
+    int cursorWindowSize,
+    bool singleInstance = true}) {
+  final options = OpenDatabaseOptions(
       version: version,
       onConfigure: onConfigure,
       onCreate: onCreate,