+++ sqflite/lib/src/sqflite_impl.dart
@@ -3,9 +3,6 @@ import 'dart:core';
 
 import 'services_impl.dart';
 
-export 'package:sqflite/src/collection_utils.dart';
-export 'package:sqflite/src/utils.dart';
-
 /// Sqflite channel name
 const String channelName = 'com.tekartik.sqflite';
 
@@ -13,7 +10,7 @@ const String channelName = 'com.tekartik.sqflite';
 const MethodChannel channel = MethodChannel(channelName);
 
 /// Temp flag to test concurrent reads
-final bool supportsConcurrency = false;
+final supportsConcurrency = false;
 
 /// Invoke a native method
 Future<T> invokeMethod<T>(String method, [dynamic arguments]) =>